tests PublisherrRestServiceIntegration

+++ src/main/kotlin/ar/edu/unq/desapp/grupoL/criptop2p/persistence/TransaccionRepository.kt
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository
 
 @Configuration
 @Repository
-interface TransaccionRepository: CrudRepository<Transaccion?, Int?> {
+interface TransaccionRepository: CrudRepository<Transaccion?, Long?> {
     override fun findAll(): List<Transaccion>
     fun save(transaccion: Transaccion): Transaccion
